Split etcdop pkg to more files
