+++ internal/pkg/service/common/etcdop/op/result.go
@@ -1,0 +1,24 @@
+package op
+
+import (
+	"go.etcd.io/etcd/api/v3/mvccpb"
+)
+
+// KeyValue - operation result.
+type KeyValue = mvccpb.KeyValue
+
+type KeyValuesT[T any] []KeyValueT[T]
+
+// KeyValueT - typed operation result.
+type KeyValueT[T any] struct {
+	Value T
+	KV    *KeyValue
+}
+
+func (kvs KeyValuesT[T]) Values() (out []T) {
+	out = make([]T, len(kvs))
+	for i, kv := range kvs {
+		out[i] = kv.Value
+	}
+	return out
+}
